refactor: extract neural scoring logic and tests into separate files
Split neural/mod.rs (634→186 lines) into:
- scoring.rs (304 lines): score_text, score_paths, score_segments, forward_all + helpers
- tests.rs (152 lines): all unit tests
